Day13 Finish Part One

+++ day13/src/main/java/com/gabrieltelles/adventofcode2024/day13/Machine.java
@@ -1,4 +1,4 @@
 package com.gabrieltelles.adventofcode2024.day13;
 
-public record Machine(int ax, int ay, int bx, int by, int prizeX, int prizeY) {
+public record Machine(long ax, long ay, long bx, long by, long prizeX, long prizeY) {
 }
